% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetsboard.R
\name{dataframe_to_reactflow_node_data}
\alias{dataframe_to_reactflow_node_data}
\title{dataframe_to_reactflow_node_data}
\usage{
dataframe_to_reactflow_node_data(dataframe)
}
\description{
dataframe_to_reactflow_node_data
}
