% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f}
\alias{f}
\title{F string}
\usage{
f(
  ...,
  .sep = "",
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  .na = "NA",
  .null = character(),
  .comment = "#",
  .literal = FALSE,
  .transformer = identity_transformer,
  .trim = TRUE
)
}
\description{
Alias to glue::glue to resemble the Pythonic f-string.
}
\examples{
mean <- 3.5
f("the mean is {mean}")
}
\seealso{
\code{\link[glue:glue]{glue::glue()}}
}
\keyword{internal}
